[9.0][FIX]product_variant_configurator: fix create_variant_ids
In standard, when attributes are added to an already created product
template, the variant without attributes is deleted or deactivated (been
used somewhere else) automatically.
In product variant configurator module we loosing that behavior when the
creation of the variants are delayed (i.e case where no auto variants are created).
This PR aims to restore that behavior by deactivating the variant with
no attributes.