[META] Make seller_ids and tmpl_seller_ids stored
When confirming complex sales orders, calls to `product.seller_ids`
in name_get() and select_seller() result in the ORM re-calculating this fields
values for many products due to running out of cache (i believe).

This result in ~6s compute time per line being wasted on re-running the lambdas
In my case this compute function was being hit 800,000 times on a 8 line SO.
Making the field stored resolves this issue,
but obviously has some impact when initially installing the module.

There appears to be no detrimental effects to making this change.