[META] The existing compute logic doesn't work
for KFs use case, and I can't quite see how it worked before.
This change should require no force updates of computes on staging,
however a script is being run on dev-docker as we speak with this fix applied:

product = odoo.env['product.product']
products = product.search([])
print(products)
for prod in products:
    print("recomputing seller ids for product id {}".format(prod))
    product.browse(prod).compute_seller_ids()

This is required due to the compute not re-running after code change,
and it does not appear to be possible to drop a column which is a M2M, like
we would do if we wanted to  force re-compute any other field.