[FIX] product_variant_default_code: recurrent prefix
With a non automatic reference mask that has a code prefix, when that
reference mask is recomputed, we could be repeating the prefix without
noticing it. For instance, when we change an product.attribute name and
that triggers all the default codes recomputations.

In a product template with variants (i.e.: Conference Chair)

- You'll have a reference mask ([Legs])
- Add a prefix: cc

--> The prefix is duplicated in the reference mask: cccc[Legs]

- Now you can delete de prefix manually from the mask and the resulting mask
  will be right: cc[Legs]
- Now change the name of the attribute Legs to something else.

--> The mask computation is retriggered and we get a duplicated prefix again cccc[Legs].

An extra issue would be that the mask isn't recomputed with the new attribute
name, although that would need some extra work.

TT48322