[FIX] puchase_variant_configurator: Avoid error when similar sellers found
The standard code for selecting the seller (product.supplierinfo)
performs a sorted over sequence, min_qty, price and ending with id.
When calling it from this module with a virtual record, as we do not
having yet the variant created, .id gets NewId type records (even if
the supplierinfo exists), so we get the error:

TypeError: '<' not supported between instances of 'NewId' and 'NewId'

For avoiding this, let's remove the id from the sorting criteria as
a workaround.

TT55224